refactored code to add logger
