+++ fintech-auth/cmd/api/main.go
@@ -10,14 +10,13 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-
 func main() {
-
+	logger := configs.Getlogger()
 	conf := configs.NewConfig()
 	db := configs.ConnectToDB(conf)
 	engine := gin.Default()
 	repo := repo.NewUser(db)
-	usecase := usecases.NewUserUseCase(repo)
+	usecase := usecases.NewUserUseCase(repo, logger)
 	controllers.NewHandler(engine, usecase)
 
 	if err := engine.Run(conf.String("port")); err != nil {
