+++ fintech-auth/internal/controllers/create.go
@@ -1,40 +1,29 @@
 package controllers
 
 import (
+	"errors"
 	pkg "fintechGo/internal/pkg"
 	"fintechGo/internal/types"
-	services "fintechGo/internal/usecases/interfaces"
+
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
-type Userhandler struct {
-	userUsercase services.UserInterface
-}
-
-func NewHandler(ctx *gin.Engine, user services.UserInterface) {
-	handler := &Userhandler{
-		userUsercase: user,
-	}
-	ctx.POST("/", handler.CreateUser)
-
-}
-
 func (u *Userhandler) CreateUser(ctx *gin.Context) {
 	var users *types.AuthUser
 	if err := ctx.ShouldBindJSON(&users); err != nil {
-		pkg.ErrorResponse(ctx, http.StatusNotFound, err)
+		pkg.ErrorResponse(ctx, http.StatusNotFound, errors.New("bad payload"))
 		return
 	}
 	jwt, err := u.userUsercase.CreateUser(users)
 	if err != nil {
-		pkg.ErrorResponse(ctx, http.StatusInternalServerError, err)
+		pkg.ErrorResponse(ctx, http.StatusInternalServerError, errors.New("error encountered"))
 		return
 	}
 	ctx.JSON(200, gin.H{
+		"password": users.Password,
 		"message":  "success",
 		"token":    jwt,
 	})
-
 }
