+++ fintech-auth/internal/controllers/routes.go
@@ -1,0 +1,19 @@
+package controllers
+
+import (
+	services "fintechGo/internal/usecases/interfaces"
+
+	"github.com/gin-gonic/gin"
+)
+
+type Userhandler struct {
+	userUsercase services.UserInterface
+}
+
+func NewHandler(ctx *gin.Engine, user services.UserInterface) {
+	handler := &Userhandler{
+		userUsercase: user,
+	}
+	ctx.POST("/", handler.CreateUser)
+
+}
