+++ fintech-auth/internal/usecases/create_user.go
@@ -5,10 +5,13 @@ import (
 	"fintechGo/internal/repo/interfaces"
 	"fintechGo/internal/types"
 	services "fintechGo/internal/usecases/interfaces"
+
+	"github.com/rs/zerolog"
 )
 
 type UserUsecase struct {
 	userRepo interfaces.AuthRepo
+	logger   *zerolog.Logger
 }
 
 // CreateUser implements interfaces.UserInterface
@@ -16,14 +19,18 @@ func (r *UserUsecase) CreateUser(user *types.AuthUser) (map[string]string, error
 
 	password, err := user.HashPassword(user.Password)
 	if err != nil {
+		r.logger.Error().Err(err).Send()
 		return nil, err
 	}
+
 	user.Password = password
 	if err := r.userRepo.CreateUser(user); err != nil {
+		r.logger.Error().Send()
 		return nil, err
 	}
 	data, err := pkg.CreateJWT(user.Email)
 	if err != nil {
+		r.logger.Error().Send()
 		return nil, err
 	}
 	return data, nil
@@ -34,8 +41,9 @@ func (*UserUsecase) FindUserByData(data string) (*types.AuthUser, error) {
 	panic("unimplemented")
 }
 
-func NewUserUseCase(user interfaces.AuthRepo) services.UserInterface {
+func NewUserUseCase(user interfaces.AuthRepo, logger *zerolog.Logger) services.UserInterface {
 	return &UserUsecase{
 		userRepo: user,
+		logger:   logger,
 	}
 }
